bluetooth: audio: Add API to get MCS service ATT handles
This is needed for upper tester.
